+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade Ruby from 2.5.1 to 2.5.3
 - Update gems to the latest versions
 - Bugfix on support list, now transferred memberships show user's details correctly
+- Order displayed membership offers by price, highest to lowest
 
 ### Removed
 - Nothing significant in this release